requirements.txt
pyspark>=3.5.0
pandas
numpy
pyarrow
boto3>=1.34.0
requests
matplotlib
seaborn
jupyter
pyyaml>=6.0
python-dotenv>=1.0.1
